+++ App/HttpServer/CMakeLists.txt
@@ -1,13 +1,13 @@
 cmake_minimum_required(VERSION 3.10)
 project(httpServer)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
-set(BOOST_ROOT /usr/local/include/boost)
+add_definitions(-DBOOST_ERROR_CODE_HEADER_ONLY)
 
-find_package(Boost 1.76.0 REQUIRED)
 set(Boost_USE_STATIC_LIBS        ON)
 set(Boost_USE_MULTITHREADED      ON)
+find_package(Boost COMPONENTS ALL REQUIRED)
 
 if(Boost_FOUND)
     include_directories(${Boost_INCLUDE_DIR} httpServer httpServer/managers ../httpClient/httpClient)
@@ -16,8 +16,7 @@ if(Boost_FOUND)
 #    add_library(SoundManager httpServer/managers/soundManager.cpp httpServer/managers/serializer.cpp)
 
     add_library(httpServer httpServer/httpServer.cpp httpServer/httpConnection.cpp httpServer/router.cpp
-                            httpServer/parser.cpp httpServer/reply.cpp)
-#    target_link_libraries(httpServer)
+                            httpServer/parser.cpp httpServer/reply.cpp httpServer/mime_types.cpp)
 
     add_executable(httpServer_run main.cpp)
     target_link_libraries(httpServer_run httpServer)
