+++ App/HttpServer/httpServer/httpConnection.cpp
@@ -1,7 +1,6 @@
-#include <boost/bind.hpp>
-
-#include "router.h"
 #include "httpConnection.h"
+
+#include <boost/bind.hpp>
 
 HttpConnection::HttpConnection(boost::asio::io_service& io_service,
                        Router& handler)
@@ -9,13 +8,11 @@ HttpConnection::HttpConnection(boost::asio::io_service& io_service,
           socket_(io_service),
           router_(handler) {}
 
-boost::asio::ip::tcp::socket& HttpConnection::socket()
-{
+boost::asio::ip::tcp::socket& HttpConnection::socket() {
     return socket_;
 }
 
-void HttpConnection::start()
-{
+void HttpConnection::start() {
     socket_.async_read_some(boost::asio::buffer(buffer_),
                             strand_.wrap(
                                     boost::bind(&HttpConnection::handle_read, shared_from_this(),
@@ -24,10 +21,8 @@ void HttpConnection::start()
 }
 
 void HttpConnection::handle_read(const boost::system::error_code& e,
-                             std::size_t bytes_transferred)
-{
-    if (!e)
-    {
+                             std::size_t bytes_transferred) {
+    if (!e) {
         boost::tribool result;
         boost::tie(result, boost::tuples::ignore) = parser_.parse(
                 request_, buffer_.data(), buffer_.data() + bytes_transferred);
