+++ App/HttpServer/httpServer/mime_types.h
@@ -1,0 +1,3 @@
+#include <string>
+
+std::string extension_to_type(const std::string& extension);
